Added the Player component which pauses and plays a song
