+++ app/components/Footer.jsx
@@ -2,6 +2,8 @@ import styles from './Footer.module.css';
 
 export default function Footer() {
   return (
-		<footer className={styles.footer}><p>v0.4.0-alpha</p></footer>
+		<footer className={styles.footer}>
+			<p>v0.4.0-alpha</p>
+		</footer>
 	);
 }
