+++ app/components/Player.module.css
@@ -1,0 +1,21 @@
+.playButton {
+	background: none;
+	border: none;
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	translate: -50% -50%;
+	width: 50%;
+	height: 50%;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
+
+.playButton svg {
+	width: 100%;
+	height: 100%;
+	opacity: 0.8;
+	cursor: pointer;
+	fill: white;
+}
