+++ app/music/SongCover.module.css
@@ -15,25 +15,3 @@
 	height: inherit;
 	border-radius: inherit;
 }
-
-.playButton {
-	background: none;
-	border: none;
-	position: absolute;
-	top: 50%;
-	left: 50%;
-	translate: -50% -50%;
-	width: 50%;
-	height: 50%;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
-
-.playButton svg {
-	width: 100%;
-	height: 100%;
-	opacity: 0.8;
-	cursor: pointer;
-	fill: white;
-}
